add activity management features and refactor entities
- Add activity API endpoints and OpenAPI configuration
- Implement activity controller, service, repository, and DTOs
- Update Activity entity: rename 'name' to 'title', change benefitsCtxh to Double, remove DRAFT status
- Enhance error handling with field-level validation errors
- Remove isActive field from User entity
- Update Student CTXH days calculation to support decimal values

+++ src/main/java/com/ctxh/volunteer/module/student/entity/Student.java
@@ -119,7 +119,7 @@ public class Student extends BaseEntity {
     /**
      * Update CTXH hours
      */
-    public void updateCtxhDays(int days) {
+    public void updateCtxhDays(double days) {
         if (days > 0) {
             this.totalCtxhDays += days;
         }
